Back/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pandas
numpy
tensorflow
